+++ src/commands/helps/doc.rs
@@ -24,8 +24,8 @@ Usage:
 Options:
     --open                open the docs in a browser after the operation
     --no-deps             don't build documentation for dependencies
-    --document-private-items
-                          document private items
+    --document-private-units
+                          document private units
     --target-dir <dir>    directory for all generated artifacts and intermediate files
 
 Use 'orbit help doc' to read more about the command."#;